Using et2_dialog for popup

+++ notifications/js/notificationajaxpopup.js
@@ -366,8 +366,13 @@
 		jQuery('.egwpopup_toggle').click(function(){window.app.notifications.toggle();});
 		jQuery('#egwpopup_fw_notifications').click(function(){window.app.notifications.toggle();});
 		jQuery(".egwpopup_deleteall", '#egwpopup').click(function(){
-	        if ( confirm(egw.lang('Are you sure you want to delete all notifications?')) )
-			window.app.notifications.delete_all()
+			et2_dialog.show_dialog( function(_button){
+					if (_button == 2) window.app.notifications.delete_all();
+				}, 
+				egw.lang('Are you sure you want to delete all notifications?'), 
+				egw.lang('Delete notifications'), 
+				null, et2_dialog.BUTTON_YES_NO, et2_dialog.WARNING_MESSAGE, undefined, egw
+			);
 		});
 		jQuery(".egwpopup_seenall", '#egwpopup').click(function(){window.app.notifications.mark_all_seen()});
 	});
